Factor out code common to measures and analyse charts
