+++ apps/base.py
@@ -1,5 +1,7 @@
 """Callbacks that apply to all pages
 """
+import re
+
 from data import get_test_code_to_name_map
 from data import get_entity_label_to_id_map
 
@@ -122,3 +124,13 @@ def toggle_entity_id_list_from_click_data(click_data, entity_ids):
         else:
             entity_ids.remove(entity_id)
     return entity_ids
+
+
+def filter_entity_ids_for_type(entity_type, entity_ids):
+    if entity_type == "ccg_id":
+        rx = r"^[0-9]{2}[A-Z]{1}$"
+    elif entity_type == "practice_id":
+        rx = r"^[A-Z]{1}[0-9]{5}$"
+    else:
+        rx = r"^[a-z]+$"
+    return [x for x in entity_ids if re.match(rx, x)]
